move main.py and session.py to internal folder
