+++ src/goose/_internal/cli/main.py
@@ -6,8 +6,8 @@ import click
 from rich import print
 from ruamel.yaml import YAML
 
-from ..config import SESSIONS_PATH
-from goose.cli.session import Session
+from ...config import SESSIONS_PATH
+from .session.session import Session
 from goose.utils import load_plugins
 from goose.utils.session_file import list_sorted_session_files
